refactor(clientSSLContext): created different file
